軽く実装してみた<-core_priest
