+++ fireplaceAharaLab/fireplace/cards/core/__init__.py
@@ -2,5 +2,6 @@ from .druid import *
 from .neutral import *
 from .hunter import *
 from .mage import *
+from .priest import *
 from .warrior import *
 from .warlock import *
